+++ Thunderstore_Stuff/CHANGELOG.md
@@ -1,3 +1,11 @@
+# Version 1.2.0
+- Refactored some code
+- Added UserID and LastChange tracking
+	- UserID is used for comparison between current data and old data now since the name changing has gotten out of hand
+	- LastChange will display on wether the entry has been changed or not
+- Added autogeneration of the necessary UserData Folders 
+# Version 1.1.2
+- Squashed some bugs
 # Version 1.1.1
 - Changed listseparator from ; to ,
 - Added additional detection for color codes with the format of "<color=#000000>"